fix lint and replace metrics enum with a list of strings
