--- ingestr/src/linkedin_ads/metrics_dimenison_enum.py
@@ -1,120 +1,0 @@
-from enum import Enum
-
-
-class Dimension(Enum):
-    campaign = "campaign"
-    creative = "creative"
-    account = "account"
-
-
-class TimeGranularity(Enum):
-    daily = "DAILY"
-    monthly = "MONTHLY"
-
-
-class Metric(Enum):
-    actionClicks = "actionClicks"
-    adUnitClicks = "adUnitClicks"
-    approximateMemberReach = "approximateMemberReach"
-    averageDwellTime = "averageDwellTime"
-    audiencePenetration = "audiencePenetration"
-    cardClicks = "cardClicks"
-    cardImpressions = "cardImpressions"
-    clicks = "clicks"
-    commentLikes = "commentLikes"
-    comments = "comments"
-    companyPageClicks = "companyPageClicks"
-    conversionValueInLocalCurrency = "conversionValueInLocalCurrency"
-    costInLocalCurrency = "costInLocalCurrency"
-    costInUsd = "costInUsd"
-    costPerQualifiedLead = "costPerQualifiedLead"
-    dateRange = "dateRange"
-    documentCompletions = "documentCompletions"
-    documentFirstQuartileCompletions = "documentFirstQuartileCompletions"
-    documentMidpointCompletions = "documentMidpointCompletions"
-    documentThirdQuartileCompletions = "documentThirdQuartileCompletions"
-    downloadClicks = "downloadClicks"
-    externalWebsiteConversions = "externalWebsiteConversions"
-    externalWebsitePostClickConversions = "externalWebsitePostClickConversions"
-    externalWebsitePostViewConversions = "externalWebsitePostViewConversions"
-    follows = "follows"
-    fullScreenPlays = "fullScreenPlays"
-    headlineClicks = "headlineClicks"
-    headlineImpressions = "headlineImpressions"
-    impressions = "impressions"
-    jobApplications = "jobApplications"
-    jobApplyClicks = "jobApplyClicks"
-    landingPageClicks = "landingPageClicks"
-    leadGenerationMailContactInfoShares = "leadGenerationMailContactInfoShares"
-    leadGenerationMailInterestedClicks = "leadGenerationMailInterestedClicks"
-    likes = "likes"
-    oneClickLeadFormOpens = "oneClickLeadFormOpens"
-    oneClickLeads = "oneClickLeads"
-    opens = "opens"
-    otherEngagements = "otherEngagements"
-    pivotValues = "pivotValues"
-    postClickJobApplications = "postClickJobApplications"
-    postClickJobApplyClicks = "postClickJobApplyClicks"
-    postClickRegistrations = "postClickRegistrations"
-    postViewJobApplications = "postViewJobApplications"
-    postViewJobApplyClicks = "postViewJobApplyClicks"
-    postViewRegistrations = "postViewRegistrations"
-    qualifiedLeads = "qualifiedLeads"
-    reactions = "reactions"
-    registrations = "registrations"
-    sends = "sends"
-    shares = "shares"
-    subscriptionClicks = "subscriptionClicks"
-    talentLeads = "talentLeads"
-    textUrlClicks = "textUrlClicks"
-    totalEngagements = "totalEngagements"
-    validWorkEmailLeads = "validWorkEmailLeads"
-    videoCompletions = "videoCompletions"
-    videoFirstQuartileCompletions = "videoFirstQuartileCompletions"
-    videoMidpointCompletions = "videoMidpointCompletions"
-    videoStarts = "videoStarts"
-    videoThirdQuartileCompletions = "videoThirdQuartileCompletions"
-    videoViews = "videoViews"
-    viralCardClicks = "viralCardClicks"
-    viralCardImpressions = "viralCardImpressions"
-    viralClicks = "viralClicks"
-    viralCommentLikes = "viralCommentLikes"
-    viralComments = "viralComments"
-    viralCompanyPageClicks = "viralCompanyPageClicks"
-    viralDocumentCompletions = "viralDocumentCompletions"
-    viralDocumentFirstQuartileCompletions = "viralDocumentFirstQuartileCompletions"
-    viralDocumentMidpointCompletions = "viralDocumentMidpointCompletions"
-    viralDocumentThirdQuartileCompletions = "viralDocumentThirdQuartileCompletions"
-    viralDownloadClicks = "viralDownloadClicks"
-    viralExternalWebsiteConversions = "viralExternalWebsiteConversions"
-    viralExternalWebsitePostClickConversions = (
-        "viralExternalWebsitePostClickConversions"
-    )
-    viralExternalWebsitePostViewConversions = "viralExternalWebsitePostViewConversions"
-    viralFollows = "viralFollows"
-    viralFullScreenPlays = "viralFullScreenPlays"
-    viralImpressions = "viralImpressions"
-    viralJobApplications = "viralJobApplications"
-    viralJobApplyClicks = "viralJobApplyClicks"
-    viralLandingPageClicks = "viralLandingPageClicks"
-    viralLikes = "viralLikes"
-    viralOneClickLeadFormOpens = "viralOneClickLeadFormOpens"
-    viralOneClickLeads = "viralOneClickLeads"
-    viralOtherEngagements = "viralOtherEngagements"
-    viralPostClickJobApplications = "viralPostClickJobApplications"
-    viralPostClickJobApplyClicks = "viralPostClickJobApplyClicks"
-    viralPostClickRegistrations = "viralPostClickRegistrations"
-    viralPostViewJobApplications = "viralPostViewJobApplications"
-    viralPostViewJobApplyClicks = "viralPostViewJobApplyClicks"
-    viralPostViewRegistrations = "viralPostViewRegistrations"
-    viralReactions = "viralReactions"
-    viralRegistrations = "viralRegistrations"
-    viralShares = "viralShares"
-    viralSubscriptionClicks = "viralSubscriptionClicks"
-    viralTotalEngagements = "viralTotalEngagements"
-    viralVideoCompletions = "viralVideoCompletions"
-    viralVideoFirstQuartileCompletions = "viralVideoFirstQuartileCompletions"
-    viralVideoMidpointCompletions = "viralVideoMidpointCompletions"
-    viralVideoStarts = "viralVideoStarts"
-    viralVideoThirdQuartileCompletions = "viralVideoThirdQuartileCompletions"
-    viralVideoViews = "viralVideoViews"
